Relocated key loader functions to shared crypto library
