+++ src/bsl/crypto/KeyLoader.h
@@ -1,0 +1,56 @@
+/*
+ * Copyright (c) 2025-2026 The Johns Hopkins University Applied Physics
+ * Laboratory LLC.
+ *
+ * This file is part of the Bundle Protocol Security Library (BSL).
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * This work was performed for the Jet Propulsion Laboratory, California
+ * Institute of Technology, sponsored by the United States Government under
+ * the prime contract 80NM0018D0004 between the Caltech and NASA under
+ * subcontract 1700763.
+ */
+/** @file
+ * @ingroup crypto
+ * Common functions to load keys from file.
+ */
+#ifndef BSL_CRYPTO_KEYLOADER_H_
+#define BSL_CRYPTO_KEYLOADER_H_
+
+#include "KeyStore.h"
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+/** Load keys from file.
+ *
+ * @param[in] file_path path to JSON file with JWK Set or CBOR file with @c COSE_KeySet
+ * @return 0 if successful.
+ */
+int BSL_Crypto_KeyLoader_LoadFile(const char *file_path);
+
+/** @warning Exposed only for testing.
+ * @param infd The file descriptor to read from.
+ */
+int BSL_Crypto_KeyLoader_LoadJwkSet(int infd);
+
+/** @warning Exposed only for testing.
+ * @param infd The file descriptor to read from.
+ */
+int BSL_Crypto_KeyLoader_LoadCoseKeySet(int infd);
+
+#ifdef __cplusplus
+} // extern C
+#endif
+
+#endif /* BSL_CRYPTO_KEYLOADER_H_ */
